+++ web/assets/js/site.js
@@ -13,11 +13,11 @@ var mapBbox = [mapBounds.top, mapBounds.lef, mapBounds.bot, mapBounds.rig];
 // map open location
 var mapCentre = [50.8424, 0.4676], mapZoom = 15;
 // map base layer
-var actTileLayer = 'osmstd';
+var defTileLayer = 'osmstd', actTileLayer = defTileLayer;
 // tab to open
-var actTab = 'home';
+var defTab = 'home', actTab = defTab;
 
-// get safari to scroll tour tab containing iframe
+// hack to get safari to scroll iframe
 if (navigator.userAgent.indexOf("Safari") > -1) {
 	$("#tour .sidebar-body").css("-webkit-overflow-scrolling", "touch");
 	$("#tour .sidebar-body").css("overflow", "auto");
@@ -35,11 +35,13 @@ $(document).on('click', 'a[href*="#goto"]', function (e) {
 	e.preventDefault();
 });
 
-// get current sidebar-tab
 $('.sidebar-tabs').click(function () {
-	if (actTab !== $('.sidebar-pane.active').attr('id')) actTab = $('.sidebar-pane.active').attr('id');
+	// get current sidebar-tab
+	actTab = $('.sidebar-pane.active').attr('id');
 	// resize links on minimap
 	if (actTab === 'home') setTimeout(function () { $('map').imageMapResize(); }, 500);
+	// hack to stop iframe freezing on firefox android
+	else if (actTab === 'tour' && L.Browser.touch) $('#tourList').trigger('change');
 });
 
 $(document).ready(function () {
@@ -48,7 +50,7 @@ $(document).ready(function () {
 	setTimeout(function () { $('map').imageMapResize(); }, 500);
 	// set collapsible accordion for opening hours
 	map.on('popupopen', function () {
-		// delay required when switching popups immediately
+		// delay required when switching directly to another popup
 		setTimeout(function () {
 			// opening-hours accordion
 			$('#accordOh').accordion({
@@ -57,12 +59,23 @@ $(document).ready(function () {
 				active: false,
 				animate: 100
 			});
+			// fhrs api for food hygine ratings
+			if ($('#fhrsLink').length) {
+				$.ajax({
+					url: 'http://api.ratings.food.gov.uk/establishments/' + $('#fhrsLink').text(),
+					headers: { 'x-api-version': 2 },
+					dataType: 'json',
+					success: function (result) {
+						$('#fhrsLink').html('<img src="assets/img/fhrs/' + result.RatingKey + '.jpg">');
+					}
+				});
+			}
 			// wikimedia api for image attribution
 			if ($('#wikiImg').length) {
 				var img = $('#wikiImg').attr('href');
 				img = img.split('File:');
 				$.ajax({
-					url: '//commons.wikipedia.org/w/api.php',
+					url: 'http://commons.wikipedia.org/w/api.php',
 					data: { action: 'query', prop: 'imageinfo', iiprop: 'extmetadata',  titles: 'File:' + img[1], format: 'json' },
 					dataType: 'jsonp',
 					success: function (result) {
@@ -74,7 +87,7 @@ $(document).ready(function () {
 					}
 				});
 			}
-		}, 250);
+		}, 200);
 	});
 });
 
@@ -120,7 +133,7 @@ function walkPoint(e) {
 }
 function improveMap(e) {
 	// create a note on osm.org
-	window.open('http://www.openstreetmap.org/note/new#map=' + map.getZoom() + '/' + e.latlng.lat + '/' + e.latlng.lng, '_blank');
+	window.open('https://www.openstreetmap.org/note/new#map=' + map.getZoom() + '/' + e.latlng.lat + '/' + e.latlng.lng, '_blank');
 }
 
 // navigation controls for historic tour
@@ -306,14 +319,16 @@ L.easyButton({
 			var uri = URI(window.location.href);
 			var selectedPois = '', walkCoords = '', tourPage, openNow;
 			var walkWayp = routingControl.getWaypoints();
+			if (actTab === defTab) actTab = undefined;
+			if (actTileLayer === defTileLayer) actTileLayer = undefined;
 			if (walkWayp[1].latLng !== null) {
 				for (var c in walkWayp) {
 					walkCoords += Math.round(walkWayp[c].latLng.lat * 100000) / 100000 + 'x' + Math.round(walkWayp[c].latLng.lng * 100000) / 100000 + '_';
 				}
 			}
 			walkCoords = walkCoords ? walkCoords = walkCoords.slice(0, -1) : undefined;
-			if (actTab === 'tour') tourPage = $('#tourList option:selected').val();
-			if ($('input.opennow').is(':checked')) openNow = 1;
+			if (actTab === 'tour' && $('#tourList option:selected').val() > 0) tourPage = $('#tourList option:selected').val();
+			if ($('.opennow input').is(':checked')) openNow = 1;
 			$('#pois input.poi-checkbox:checked').each(function (i, element) { selectedPois += element.dataset.key + '-'; });
 			selectedPois = selectedPois ? selectedPois = selectedPois.slice(0, -1) : undefined;
 			// M = basemap, T = tab, U = tour page, O = opennow, P = grouped pois, I = single poi, W = walkpoints
@@ -359,7 +374,7 @@ var routingControl = L.Routing.control({
 	collapsible: false,
 	fitSelectedRoutes: false,
 	reverseWaypoints: true,
-	routeWhileDragging: true,
+	routeWhileDragging: false,
 	router: new L.Routing.mapbox(mapboxKey, {
 		profile: 'mapbox.walking',
 		alternatives: false
@@ -410,11 +425,10 @@ function populate_tabs() {
 				clear_map();
 				$('a[href="#pois"]').click();
 				$('#pois input[id="' + catSplit[1] + '"]').prop('checked', true);
-				// scroll to and highlight checkbox
+				// scroll to checkbox
 				if ($(window).width() >= 768) {
 					$('#pois .sidebar-body').scrollTop(0);
 					$('#pois .sidebar-body').scrollTop($('#pois input[id="' + catSplit[1] + '"]').position().top - 50);
-					$('#pois input[id="' + catSplit[1] + '"]').parent().parent().parent().effect('highlight', {}, 3000);
 				}
 				setting_changed();
 				$('#autocomplete').val('');
@@ -438,10 +452,10 @@ function populate_tabs() {
 			if (pois[poi].catName === categoryList[c]) {
 				if (t % 2 === 0) checkboxContent += '<tr>';
 				checkboxContent += L.Util.template(
-					'<td style="overflow:hidden; white-space:nowrap;"> \
+					'<td> \
 					<div class="poi-checkbox"> \
 						<label title="{name}"> \
-							<img style="width:28px;" src="assets/img/icons/{icon}.png"></img> \
+							<img src="assets/img/icons/{icon}.png"></img> \
 							<input type="checkbox" class="poi-checkbox" id="{name}" data-key="{key}" onclick="setting_changed(&#39;{key}&#39;)"><span>{name}</span> \
 						</label> \
 					</div> \
@@ -453,156 +467,11 @@ function populate_tabs() {
 			}
 		}
 		checkboxContent += '</table>';
-		checkboxContent += '<div class="anchor"><a href="#gototoppois">| <i class="fa fa-arrow-up"></i> |</a></div>';
+		checkboxContent += '<div class="anchor"><a href="#gototoppois" title="Back to top">| <i class="fa fa-arrow-up"></i> |</a></div>';
 	}
 	$('#pois .sidebar-body').append(checkboxContent);
 }
 populate_tabs();
-
-// https://github.com/kartenkarsten/leaflet-layer-overpass
-var spinner = 0, markerId;
-function callback(data) {
-	var type, markerPopup;
-	// set marker popup dimensions on screensize
-	var customOptions = ($(window).width() >= 768) ? { maxWidth: 350 } : { maxWidth: 250, className: 'popup-mobile' };
-	customOptions.autoPanPaddingTopLeft = (($(window).width() >= 768) && rLookup) ? [sidebar.width()+50,5] : [30,5];
-	customOptions.autoPanPaddingBottomRight = [5,75];
-	customOptions.minWidth = 205;
-	customOptions.closeButton = false;
-	if (spinner > 0) spinner--;
-	if (spinner === 0) $('#spinner').hide();
-	for (var c in data.elements) {
-		var e = data.elements[c];
-		if (e.id in this.instance._ids) continue;
-		this.instance._ids[e.id] = true;
-		var pos = (e.type === 'node') ? new L.LatLng(e.lat, e.lon) : new L.LatLng(e.center.lat, e.center.lon);
-		type = undefined;
-		if (siteDebug) console.debug(e);
-		if (e.tags.amenity) {
-			if (!type) type = e.tags.amenity;
-			// Group similar pois
-			if (type === 'arts_centre') type = 'attraction';
-			if (type === 'nightclub') type = 'bar';
-			if (type === 'college') type = 'school';
-			if (type === 'school' && e.tags.name === undefined) type = undefined;
-			if (type === 'retirement_home') type = 'social_facility';
-			if (type === 'post_office') type = 'post_box';
-			// Hide non-public parking
-			if (type === 'parking') {
-				if (e.tags.access === 'private' || e.tags.access === 'permissive') type = undefined;
-			}
-			if (type === 'animal_boarding') type = 'animal_shelter';
-		}
-		if (e.tags.historic) {
-			if (!type) type = 'historic';
-		}
-		if (e.tags.man_made) {
-			if (!type) type = e.tags.man_made;
-		}
-		if (e.tags.shop) {
-			if (!type) type = e.tags.shop;
-			// Group similar pois
-			if (type === 'deli') type = 'butcher';
-			if (type === 'e-cigarette') type = 'tobacco';
-			if (type === 'hardware') type = 'doityourself';
-			if (type === 'boutique') type = 'clothes';
-			if (type === 'window_blind') type = 'curtain';
-			if (type === 'laundry') type = 'dry_cleaning';
-			if (type === 'garden_centre') type = 'florist';
-			if (type === 'tyres') type = 'car_repair';
-			if (type === 'hearing_aids') type = 'mobility';
-			if (type === 'interior_decoration' || type === 'bathroom_furnishing' || type === 'kitchen') type = 'houseware';
-		}
-		if (e.tags.tourism) {
-			if (!type) type = e.tags.tourism;
-		}
-		if (e.tags.landuse) {
-			if (!type) type = e.tags.landuse;
-			// Hide non-public grounds
-			if (type === 'recreation_ground') {
-				if (e.tags.access === 'private') type = undefined;
-			}
-		}
-		if (e.tags.leisure) {
-			if (!type) type = e.tags.leisure;
-			if (type === 'common') type = 'park';
-			if (type === 'swimming_pool') {
-				if (e.tags.access === 'private') type = undefined;
-			}
-		}
-		if (e.tags.emergency) {
-			if (!type) type = e.tags.emergency;
-		}
-		if (e.tags.office) {
-			if (!type) type = e.tags.office;
-			if (type === 'financial') type = 'accountant';
-		}
-		if (e.tags.healthcare) {
-			if (!type) type = 'healthcare';
-		}
-		if (e.tags.listed_status) {
-			if (!type || type === 'shelter' || type === 'company') type = 'listed_status';
-		}
-		if (e.tags.image) {
-			if (!type) type = 'image';
-		}
-		var poi = pois[type];
-		var iconName = poi ? poi.iconName : '000blank';
-		var markerIcon = L.icon({
-			iconUrl: 'assets/img/icons/' + iconName + '.png',
-			iconSize: [32, 37],
-			iconAnchor: [16, 35],
-			shadowUrl: 'assets/img/icons/000shadow.png',
-			shadowAnchor: [16, 27],
-			popupAnchor: [0, -27]
-		});
-		var marker = L.marker(pos, {
-			icon: markerIcon,
-			keyboard: false,
-			riseOnHover: true
-		});
-		var name = '&hellip;';
-		if (e.tags.name) name = e.tags.name;
-		else if (e.tags.ref) name = e.tags.ref;
-		else if (poi) name = poi.name;
-		else if (e.tags.building && e.tags.building !== 'yes') {
-			name = e.tags.building;
-			name = name.replace(/\w\S*/g, function(txt){return txt.charAt(0).toUpperCase() + txt.substr(1);});
-		}
-		else if (e.tags['addr:housename']) name = e.tags['addr:housename'];
-		// show a tooltip on mouse hover
-		if (name) marker.bindTooltip(name, { direction: 'left', offset: [-15, -2] }).openTooltip();
-		// check if already defined poi
-		if (poi) {
-			// create pop-up
-			markerPopup = poi.tagParser ? poi.tagParser(e, poi.name) : generic_poi_parser(e, poi.name);
-			// set width of popup on screensize
-			// show pop-up
-			marker.bindPopup(markerPopup, customOptions);
-			// check if coming from reverselookup
-			if (rLookup) {
-				marker.addTo(this.instance).openPopup();
-				markerId = e.type + '_' + e.id;
-				rLookup = false;
-			}
-			else {
-				// display only open facilities on opennow checkbox
-				if ($('input.opennow').is(':checked')) {
-					if (state) marker.addTo(this.instance);
-				}
-				else marker.addTo(this.instance);
-			}
-		}
-		else if (rLookup) {
-			// use generic marker
-			markerPopup = generic_poi_parser(e, name);
-			marker.bindPopup(markerPopup, customOptions);
-			marker.addTo(this.instance).openPopup();
-			markerId = e.type + '_' + e.id;
-			rLookup = false;
-		}
-	}
-}
 
 // clear poi marker layers
 function clear_map() {
@@ -615,6 +484,7 @@ function clear_map() {
 function setting_changed(newcheckbox) {
 	// limit number of active checkboxes
 	if ($('input.poi-checkbox:checked').length <= 3) {
+		//$('[data-key=' + newcheckbox + ']').parent().parent().parent().css('background-color', 'rgba(255,255,255,0.6)');
 		// remove old poi markers
 		iconLayer.clearLayers();
 		markerId = undefined;
@@ -635,12 +505,25 @@ function setting_changed(newcheckbox) {
 		}
 	}
 	else $('[data-key=' + newcheckbox + ']').prop('checked', false);
+	$('input.poi-checkbox').trigger('change');
 }
+// checkbox highlight
+$('input.poi-checkbox').change(function () {
+	if ($(this).prop('checked') === true) $(this).parent().parent().parent().css('background-color', 'rgba(255,255,255,0.6)');
+	else $(this).parent().parent().parent().css('background-color', '');
+});
 
 // suggested walk waypoints
 function suggWalk(walkName) {
 	clear_map();
 	switch(walkName) {
+		case 'wwh':
+			routingControl.setWaypoints([
+				[50.83567, 0.45892],
+				[50.83701, 0.47363]
+			]);
+			map.flyTo([50.8364, 0.4664], 16);
+			break;
 		case 'tmt':
 			routingControl.setWaypoints([
 				[50.84059, 0.49121],
@@ -651,6 +534,13 @@ function suggWalk(walkName) {
 			// show related information boards
 			show_overpass_layer('(node["ref"~"^TMT"](screenBbox));');
 			map.flyTo([50.8385, 0.4787], 16);
+			break;
+		case 'greenway':
+			routingControl.setWaypoints([
+				[50.85782, 0.48228],
+				[50.87008, 0.52089]
+			]);
+			map.flyTo([50.8632, 0.4938], 15);
 			break;
 		case '1066':
 			routingControl.setWaypoints([
@@ -672,13 +562,6 @@ function suggWalk(walkName) {
 			]);
 			map.flyTo([50.8474, 0.4874], 14);
 			break;
-		case 'wwh':
-			routingControl.setWaypoints([
-				[50.83567, 0.45892],
-				[50.83701, 0.47363]
-			]);
-			map.flyTo([50.8364, 0.4664], 16);
-			break;
 	}
 }
 
@@ -691,7 +574,7 @@ if (uri.hasQuery('U')) {
 		$('#tourList').val(uri.search(true).U);
 		$('#tourList').trigger('change');
 }
-if (uri.hasQuery('O')) $('input.opennow').prop('checked', uri.search(true).O);
+if (uri.hasQuery('O')) $('.opennow input').prop('checked', uri.search(true).O);
 if (uri.hasQuery('W')) {
 	var walkCoords = uri.search(true).W;
 	walkCoords = walkCoords.split("_");
@@ -705,14 +588,13 @@ if (uri.hasQuery('P')) {
 	if (selectedPois.indexOf("-") > -1) selectedPois = selectedPois.split("-");
 	if (!$.isArray(selectedPois)) {
 		$('#pois input[data-key=' + selectedPois + ']').prop('checked', true);
-		// open tab when not on mobile, scroll to and highlight checkbox
+		// open tab when not on mobile, scroll to checkbox
 		if ($(window).width() >= 768) {
 			sidebar.open(actTab);
 			$('#pois .sidebar-body').scrollTop(0);
 			// add delay after load for sidebar to animate open to allow for scroll position
 			setTimeout(function () { 
 				$('#pois .sidebar-body').scrollTop($('#pois input[data-key="' + selectedPois + '"]').position().top - 50);
-				$('#pois input[data-key=' + selectedPois + ']').parent().parent().parent().effect('highlight', {}, 3000);
 			}, 500);
 		}
 	}
@@ -741,6 +623,7 @@ if (!uri.hasQuery('W') || !uri.hasQuery('P') || !uri.hasQuery('I')) {
 tileBaseLayers[tileBaseLayer[actTileLayer].name].addTo(map);
 map.setMaxBounds([[mapBounds.top, mapBounds.lef], [mapBounds.bot, mapBounds.rig]]);
 	
+// https://github.com/kartenkarsten/leaflet-layer-overpass
 function show_overpass_layer(query) {
 	if (siteDebug) console.debug(query);
 	if (!query || query === '();') {